docs: adicionar README e reunir exercícios em um único arquivo
