+++ README.md
@@ -1,1 +1,27 @@
-# selthon06-TADS
+# Lista de Exercícios de Algoritmos
+
+Projeto desenvolvido por **Selthon** contendo **80 exercícios de lógica de programação** em **Python**, organizados por nível de dificuldade.
+
+## Estrutura
+
+- Todos os exercícios estão em **um único arquivo Python**.
+- Cada exercício está implementado em uma **função própria**.
+- Divisão dos exercícios:
+  - **Fáceis**
+  - **Médios**
+  - **Difíceis**
+  - **Dicionários**
+
+## Requisitos
+
+- **Python** (recomenda-se a versão mais recente)
+
+## Como usar
+
+1. Clone este repositório:
+
+- git clone https://github.com/selthon06/selthon06-TADS.git
+
+2. Abra o arquivo exercicios.py no editor ou terminal.
+
+3. Para executar um exercício, basta chamar a função correspondente no interpretador Python.
