--- exe1.py
@@ -1,5 +1,0 @@
-# Crie uma lista com 5 números inteiros. Imprima o tamanho da lista usando len
-
-lista = {1, 2, 3, 4, 5}
-
-print(len(lista))
