--- exe4.py
@@ -1,7 +1,0 @@
-#  Crie uma tupla com os dias da semana. Imprima apenas o terceiro dia
-
-lista_tupla = ("segunda", "terça", "quarta", "quinta", "sexta")
-
-print(lista_tupla[3])
-
-
