--- exe5.py
@@ -1,7 +1,0 @@
-# Dada a lista numeros = [1, 2, 3, 4, 5], remova o número 3 e imprima a lista
-
-lista = {1, 2, 3, 4, 5}
-
-lista.remove(3)
-
-print(lista)
